just moved the db code into server instead of exporting it
