+++ src/server/server.ts
@@ -1,7 +1,25 @@
 import express from 'express';
 // const path = require('path');
 import path from 'path';
-import connectDB from './database.ts';
+import { Client } from 'pg';
+import dotenv from 'dotenv';
+
+dotenv.config(); // crucial
+
+const connectionString: string = process.env.CONNECTION_STRING
+const client = new Client({
+  connectionString,
+});
+
+const connectDB = async () => {
+  try {
+    await client.connect();
+    console.log('Connected to ElephantSQL database');
+  } catch (error) {
+    console.error('Failed to connect to the database', error);
+  }
+};
+
 
 const app = express();
